+++ src/main/kotlin/net/glazov/data/model/users/ClientModel.kt
@@ -13,7 +13,7 @@ data class ClientModel(
     val accountNumber: String, //Номер лицевого счёта
     val address: AddressModel,
     val tariffId: String? = null, //ID подключенного тарифа
-    val balance: Double = 0.0, //Остаток средств на аккаунте
+    val balance: Float = 0.0f, //Остаток средств на аккаунте
     val accountCreationDate: Long = OffsetDateTime.now(ZoneId.systemDefault()).toEpochSecond(),
     val debitDate: Long = OffsetDateTime.now(ZoneId.systemDefault()).toEpochSecond(), //Дата списания средств
     val isAccountActive: Boolean = true,
