+++ src/main/kotlin/net/glazov/data/model/users/EmployeeModel.kt
@@ -11,6 +11,6 @@ data class EmployeeModel(
     val personId: String,
     val accountCreationDate: Long = OffsetDateTime.now(ZoneId.systemDefault()).toEpochSecond(),
     val roles: List<String>,
-    val overallRating: Double = 0.0, // sum of all ratings. Divide by ratingsCount to get average
+    val overallRating: Int = 0, // sum of all ratings (1..5). Divide by ratingsCount to get average
     val numberOfRatings: Int = 0 // number of ratings from customers
 )
